% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eANOVA.R
\name{eANOVA_dummyFunctionName}
\alias{eANOVA_dummyFunctionName}
\alias{eANOVA}
\alias{eANOVA_eta2p}
\alias{eANOVA_FStat}
\title{Effect-Size Bayes Factors for Analysis of Variance (ANOVA)}
\usage{
eANOVA_eta2p(
  eta2p,
  N,
  df_numer,
  df_denom,
  Cohen_f = c("medium", "small", "large"),
  rscale = NULL,
  nu = 3
)

eANOVA_FStat(
  FStat,
  N,
  df_numer,
  df_denom,
  Cohen_f = c("medium", "small", "large"),
  rscale = NULL,
  nu = 3
)
}
\arguments{
\item{eta2p}{observed partial coefficient of determination \eqn{{\eta}^2_p}.}

\item{FStat}{observed value of \eqn{F} statistic.}

\item{N}{number of observations.}

\item{df_numer}{numerator degrees of freedom.}

\item{df_denom}{denominator degrees of freedom.}

\item{Cohen_f}{focused-on effect size (Cohen's \eqn{f}).
Must be either numeric, or alternatively set to one of the options
\code{"medium" = .25} (default), \code{"small" = .10}, \code{"large" = .40}.}

\item{rscale}{prior scale.
If not specified, it is set internally to the recommended value.}

\item{nu}{prior degrees of freedom.
If not specified, it is set internally to the recommended value (\code{nu = 3}).}
}
\value{
An object of class \code{emBayesFactor_out} containing the following components:

\item{method}{the focussed-on analysis.}
\item{BayesFactor}{value of the Bayes factor.}
\item{accuracy}{estimate of the modulus of the absolute error.}
\item{Cohen}{the focussed-on effect size.}
\item{Cohen_size}{value of the focussed-on effect size.}
\item{rscale}{prior scale.}
\item{nu}{prior degrees of freedom.}
\item{Stat}{provided statistic.}
\item{Stat_size}{value of the provided statistic.}
\item{alternative}{the specified alternative hypothesis}
}
\description{
Computes Bayes factors (alternative over null hypothesis)
with effect-size priors for analysis of variance (ANOVA)
based on observed partial \eqn{{\eta}^2} (\eqn{{\eta}^2_p})
or the observed test statistic \eqn{F}.
}
\details{
The Bayes factor provided by \code{eANOVA} compares
the alternative hypothesis with the null hypothesis
that there is no true effect.

\code{Cohen_f} must not be negative.

If \code{rscale} is not specified, \code{nu} must be greater than 2.

If \code{Cohen_f = 0}, \code{rscale} must be explicitly specified.
}
\examples{
## Example from Klauer et al. (2023)

## Compute effect-size Bayes factor based on 'Fstat'
eANOVA_FStat(
  FStat = 16.73,
  N = 40,
  df_numer = 1,
  df_denom = 36,
  Cohen_f = "small"
)
}
\references{
Klauer, K. C., Meyer-Grant, C. G., & Kellen, D (2023).
\emph{On Bayes factors for Hypotheses Tests}. Manuscript in preparation.
}
\seealso{
\link{integrate}, \link{aov}, \link[BayesFactor]{anovaBF}
}
\author{
Constantin G. Meyer-Grant
(\email{constantin.meyer-grant@psychologie.uni-freiburg.de})
}
