% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ettest.R
\name{ettest}
\alias{ettest}
\alias{ettest_tStat}
\title{Effect-Size Bayes Factors for \eqn{t} Tests}
\usage{
ettest_tStat(
  tStat,
  N1,
  N2 = NULL,
  Cohen_d = c("medium", "small", "large"),
  rscale = NULL,
  nu = 3,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{tStat}{observed value of \eqn{t} statistic.}

\item{N1}{sample size of the first group
(or only group for one-sample tests).}

\item{N2}{sample size of the second group
(only specified for two-sample tests).}

\item{Cohen_d}{focused-on effect size (Cohen's \eqn{d}).
Must be either numeric, or alternatively set to one of the options
\code{"medium" = .50} (default), \code{"small" = .20}, \code{"large" = .80}.}

\item{rscale}{prior scale.
If not specified, it is set internally to the recommended value.}

\item{nu}{prior degrees of freedom.
If not specified, it is set internally to the recommended value (\code{nu = 3}).}

\item{alternative}{a character string specifying the alternative hypothesis.
Must be one of \code{"two.sided"} (default),
\code{"greater"}, or \code{"less"}.}
}
\value{
An object of class \code{emBayesFactor_out} containing the following components:

\item{method}{the focussed-on analysis.}
\item{BayesFactor}{value of the Bayes factor.}
\item{accuracy}{estimate of the modulus of the absolute error.}
\item{Cohen}{the focussed-on effect size.}
\item{Cohen_size}{value of the focussed-on effect size.}
\item{rscale}{prior scale.}
\item{nu}{prior degrees of freedom.}
\item{Stat}{provided statistic.}
\item{Stat_size}{value of the provided statistic.}
\item{alternative}{the specified alternative hypothesis}
}
\description{
Computes Bayes factors (alternative over null hypothesis)
with effect-size priors for one- and two-sample \eqn{t} tests
based on the observed test statistic \eqn{t}.
}
\details{
The Bayes factor provided by \code{ettest}
compares the alternative hypothesis with the null hypothesis
that there is no true effect.

Note that \code{alternative = "greater"} denotes the alternative that
the true effect is positive, whereas \code{alternative = "less"}
denotes the alternative that the true effect is negative.
Also note that in both cases (i.e., for one-sided tests)
a single truncated \eqn{t} distribution is used as prior
(see also Gronau & Wagenmakers, 2020).

\code{Cohen_d} should only be negative if \code{alternative = "less"}.

If \code{rscale} is not specified, \code{nu} must be greater than 2.

If \code{Cohen_d = 0}, the default Prior
is used as a result (Rouder et al., 2009;
see also \link[BayesFactor]{ttestBF}),
in which case \code{rscale} must be explicitly specified.
The default value specified in \link[BayesFactor]{ttestBF}
is \eqn{1/\sqrt{2}}.
}
\examples{
## Example from Klauer et al. (in press)

## Compute effect-size Bayes factor based on 'tStat'
ettest_tStat(
  tStat = 2.03,
  N1 = 80,
  Cohen_d = 0.2
)
}
\references{
Gronau, Q. F., Ly, A., & Wagenmakers, E.-J. (2020).
Informed Bayesian \emph{t}-tests.
\emph{The American Statistician}, \emph{74}(2), 137–143.

Klauer, K. C., Meyer-Grant, C. G., & Kellen, D (in press).
On Bayes Factors for Hypotheses Tests
\emph{Psychonomic Bulletin & Review}.
[PsyArXiv preprint: \href{https://doi.org/10.31234/osf.io/ykp29}{https://doi.org/10.31234/osf.io/ykp29}]

Rouder, J. N., Speckman, P. L., Sun, D., Morey, R. D., & Iverson, G.
(2009). Bayesian \emph{t}-tests for accepting and rejecting the null hypothesis.
\emph{Psychonomic Bulletin & Review}, \emph{16}(2), 225–237.
}
\seealso{
\link{integrate}, \link{t.test}, \link[BayesFactor]{ttestBF}
}
\author{
Constantin G. Meyer-Grant
(\email{constantin.meyer-grant@psychologie.uni-freiburg.de})
}
