% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mANOVA.R
\name{mANOVA_dummyFunctionName}
\alias{mANOVA_dummyFunctionName}
\alias{mANOVA}
\alias{mANOVA_eta2p}
\alias{mANOVA_FStat}
\title{Moment Bayes Factors for Analysis of Variance (ANOVA)}
\usage{
mANOVA_eta2p(
  eta2p,
  df_numer,
  df_denom,
  Cohen_f = c("medium", "small", "large"),
  nu = 5
)

mANOVA_FStat(
  FStat,
  df_numer,
  df_denom,
  Cohen_f = c("medium", "small", "large"),
  nu = 5
)
}
\arguments{
\item{eta2p}{observed partial coefficient of determination \eqn{{\eta}^2_p}.}

\item{FStat}{observed value of \eqn{F} statistic.}

\item{df_numer}{numerator degrees of freedom.}

\item{df_denom}{denominator degrees of freedom.}

\item{Cohen_f}{focused-on effect size (Cohen's \eqn{f}).
Must be either numeric, or alternatively set to one of the options
\code{"medium" = .25} (default), \code{"small" = .10}, \code{"large" = .40}.}

\item{nu}{prior degrees of freedom.
If not specified, it is set internally to the recommended value (\code{nu = 5}).}
}
\value{
An object of class \code{emBayesFactor_out} containing the following components:

\item{method}{the focussed-on analysis.}
\item{BayesFactor}{value of the Bayes factor.}
\item{accuracy}{estimate of the modulus of the absolute error.}
\item{Cohen}{the focussed-on effect size.}
\item{Cohen_size}{value of the focussed-on effect size.}
\item{nu}{prior degrees of freedom.}
\item{Stat}{provided statistic.}
\item{Stat_size}{value of the provided statistic.}
\item{alternative}{the specified alternative hypothesis}
}
\description{
Computes Bayes factors (alternative over null hypothesis)
with moment priors for analysis of variance (ANOVA)
based on observed partial \eqn{{\eta}^2} (\eqn{{\eta}^2_p})
or the observed test statistic \eqn{F}.
}
\details{
The Bayes factor provided by \code{mANOVA} compares
the alternative hypothesis with the null hypothesis
that there is no true effect.

\code{Cohen_f} must be positive.

\code{nu} must be at least 3.
}
\examples{
## Example from Klauer et al. (2023)

## Compute moment Bayes factor based on 'Fstat'
mANOVA_FStat(
  FStat = 16.73,
  df_numer = 1,
  df_denom = 36,
  Cohen_f = "small"
)
}
\references{
Klauer, K. C., Meyer-Grant, C. G., & Kellen, D (2023).
\emph{On Bayes Factors for Hypotheses Tests}.
Manuscript submitted for publication.
}
\seealso{
\link{integrate}, \link{aov}, \link[BayesFactor]{anovaBF}
}
\author{
Constantin G. Meyer-Grant
(\email{constantin.meyer-grant@psychologie.uni-freiburg.de})
}
