% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mreg.R
\name{mreg_dummyFunctionName}
\alias{mreg_dummyFunctionName}
\alias{mreg}
\alias{mreg_R2p}
\alias{mreg_FStat}
\alias{mreg_tStat}
\title{Moment Bayes Factors for (Multiple) Linear Regression}
\usage{
mreg_R2p(
  R2p,
  N,
  k,
  p,
  Cohen_f2 = c("medium", "small", "large"),
  nu = 5 + (k - p)
)

mreg_FStat(
  FStat,
  N,
  k,
  p,
  Cohen_f2 = c("medium", "small", "large"),
  nu = 5 + (k - p)
)

mreg_tStat(
  tStat,
  N,
  k,
  Cohen_f2 = c("medium", "small", "large"),
  nu = 5,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{R2p}{observed partial coefficient of determination \eqn{R^2_p}.}

\item{FStat}{observed value of \eqn{F} statistic.}

\item{tStat}{observed value of \eqn{t} statistic.}

\item{N}{number of observations.}

\item{k}{number of parameters of the full model.}

\item{p}{number of parameters of the reduced model.}

\item{Cohen_f2}{focused-on effect size (Cohen's \eqn{f^2}).
Must be either numeric, or alternatively set to one of the options
\code{"medium" = .15} (default), \code{"small" = .02}, \code{"large" = .35}.}

\item{nu}{prior degrees of freedom.
If not specified, it is set internally to the recommended value (\code{nu = 5}).}

\item{alternative}{a character string specifying the alternative hypothesis.
Must be one of \code{"two.sided"} (default),
\code{"greater"}, or \code{"less"}.}
}
\value{
An object of class \code{emBayesFactor_out} containing the following components:

\item{method}{the focussed-on analysis.}
\item{BayesFactor}{value of the Bayes factor.}
\item{accuracy}{estimate of the modulus of the absolute error.}
\item{Cohen}{the focussed-on effect size.}
\item{Cohen_size}{value of the focussed-on effect size.}
\item{nu}{prior degrees of freedom.}
\item{Stat}{provided statistic.}
\item{Stat_size}{value of the provided statistic.}
\item{alternative}{the specified alternative hypothesis}
}
\description{
Computes Bayes factors (alternative over null hypothesis)
with moment priors for (multiple) linear regression analyses
based on observed partial coefficient of determination (\eqn{R^2_p}),
the observed test statistic \eqn{F}, or the observed test statistic \eqn{t}
for testing a single partial regression coefficient.
}
\details{
The Bayes factor provided by \code{mreg} compares
the alternative hypothesis with the null hypothesis
that there is no true effect.
For \code{mreg_tStat}, \code{alternative = "greater"} denotes
the alternative that the true effect is positive,
whereas \code{alternative = "less"} denotes the alternative
that the true effect is negative.

\code{Cohen_f2} must be positive, but if \code{alternative = "less"}
is specified for \code{mreg_tStat},
the presumed relationship between predictor and criterion is negative.

\code{nu} must be at least 3.
}
\examples{
## Example from Klauer et al. (2024)

## Compute partial R^2 from the two models' R^2
R2p <- (0.7109 - 0.5670) / (1.0 - 0.5670)

## Compute moment Bayes factor based on 'R2p'
mreg_R2p(
  R2p = R2p,
  N = 175,
  p = 4,
  k = 5,
  Cohen_f2 = "large"
)
}
\references{
Klauer, K. C., Meyer-Grant, C. G., & Kellen, D (2024).
\emph{On Bayes Factors for Hypotheses Tests}.
Manuscript submitted for publication.
}
\seealso{
\link{integrate}, \link{lm}, \link[BayesFactor]{regressionBF}
}
\author{
Constantin G. Meyer-Grant
(\email{constantin.meyer-grant@psychologie.uni-freiburg.de})
}
