% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mttest.R
\name{mttest}
\alias{mttest}
\alias{mttest_tStat}
\title{Moment Bayes Factors for \eqn{t} Tests}
\usage{
mttest_tStat(
  tStat,
  N1,
  N2 = NULL,
  Cohen_d = c("medium", "small", "large"),
  nu = 5,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{tStat}{observed value of \eqn{t} statistic.}

\item{N1}{sample size of the first group
(or only group for one-sample tests).}

\item{N2}{sample size of the second group
(only specified for two-sample tests).}

\item{Cohen_d}{focused-on effect size (Cohen's \eqn{d}).
Must be either numeric, or alternatively set to one of the options
\code{"medium" = .50} (default), \code{"small" = .20}, \code{"large" = .80}.}

\item{nu}{prior degrees of freedom.
If not specified, it is set internally to the recommended value (\code{nu = 5}).}

\item{alternative}{a character string specifying the alternative hypothesis.
Must be one of \code{"two.sided"} (default),
\code{"greater"}, or \code{"less"}.}
}
\value{
An object of class \code{mBf} containing the following components:

\item{method}{the focussed-on analysis.}
\item{BayesFactor}{value of the Bayes factor.}
\item{accuracy}{estimate of the modulus of the absolute error.}
\item{Cohen}{the focussed-on effect size.}
\item{Cohen_size}{value of the focussed-on effect size.}
\item{nu}{prior degrees of freedom.}
\item{Stat}{provided statistic.}
\item{Stat_size}{value of the provided statistic.}
\item{alternative}{the specified alternative hypothesis}
}
\description{
Computes Bayes factors (alternative over null hypothesis)
with moment priors for one- and two-sample \eqn{t} tests
based on the observed test statistic \eqn{t}.
}
\details{
The Bayes factor provided by \code{mttest}
compares the alternative hypothesis with the null hypothesis
that there is no true effect.

Note that \code{alternative = "greater"} denotes the alternative that
the true effect is positive, whereas \code{alternative = "less"}
denotes the alternative that the true effect is negative.

\code{Cohen_d} should only be negative if \code{alternative = "less"} and must never be 0.

\code{nu} must be at least 3.
}
\examples{
## Example from Klauer et al. (2023)

## Compute moment Bayes factor based on 'tStat'
mttest_tStat(
  tStat = 2.03,
  N1 = 80,
  Cohen_d = 0.2
)
}
\references{
Klauer, K. C., Meyer-Grant, C. G., & Kellen, D (2023).
\emph{On Bayes Factors for Hypotheses Tests}. Manuscript in preparation.

Pramanik, S., & Johnson, V. E. (2022).
Efficient alternatives for bayesian hypothesis tests in psychology.
Advance online publication. \emph{Psychological Methods}.
}
\seealso{
\link{integrate}, \link{t.test}, \link[BayesFactor]{ttestBF}
}
\author{
Constantin G. Meyer-Grant
(\email{constantin.meyer-grant@psychologie.uni-freiburg.de})
}
